[move-only] circuit::table_layouter: Move TableLayouter, SimpleTableLayouter
